chore: create fiat currency endpoints file
